+++ maserol/figures/figureA1.py
@@ -5,17 +5,8 @@ paper.
 from tensordata.zohar import data3D as zohar
 
 from maserol.core import prepare_data
-from maserol.heatmap import plot_deviation_heatmap
-from maserol.preprocess import HIgGs
-from maserol.resample import bootstrap
+from maserol.scatterplot import plotOptimize
 
 def makeFigure():
     data = prepare_data(zohar(xarray=True, logscale=False))
-    opt_kwargs = {
-        "metric": "rtot",
-        "lrank": True,
-        "fitKa": False,
-        "ab_types": HIgGs,
-    }
-    sample_dist, ag_dist = bootstrap(data, numResample=10, norm="max", **opt_kwargs)
-    return plot_deviation_heatmap(ag_dist[0], ag_dist[1], HIgGs, data.Antigen.values)
+    return plotOptimize(data, metric="mean", lrank=True, fitKa=False, maxiter=250)
