+++ maserol/figures/figureA2.py
@@ -1,12 +1,11 @@
-"""
-This is an internal figure which does not necessarily have a place in final
-paper.
-"""
+from maserol.core import optimizeLoss, initializeParams, flattenParams
+from maserol.heatmap import plotHeatmaps
+from maserol.preprocess import prepare_data
+
 from tensordata.zohar import data3D as zohar
 
-from maserol.core import prepare_data
-from maserol.scatterplot import plot_leave_out_rec_lbound_correlation
 
 def makeFigure():
     data = prepare_data(zohar(xarray=True, logscale=False))
-    return plot_leave_out_rec_lbound_correlation(data, "FcR2B", metric="rtot")
+    x_opt, _ = optimizeLoss(data, lrank=True)
+    return plotHeatmaps(data, x_opt, lrank=True)[0]
