+++ maserol/figures/figureA3.py
@@ -1,11 +1,21 @@
-from maserol.core import optimizeLoss, initializeParams, flattenParams
-from maserol.heatmap import plotHeatmaps
-from maserol.preprocess import prepare_data
-
+"""
+This is an internal figure which does not necessarily have a place in final
+paper.
+"""
 from tensordata.zohar import data3D as zohar
 
+from maserol.core import prepare_data
+from maserol.heatmap import plot_deviation_heatmap
+from maserol.preprocess import HIgGs
+from maserol.resample import bootstrap
 
 def makeFigure():
     data = prepare_data(zohar(xarray=True, logscale=False))
-    x_opt, _ = optimizeLoss(data, lrank=True)
-    return plotHeatmaps(data, x_opt, lrank=True)[0]
+    opt_kwargs = {
+        "metric": "rtot",
+        "lrank": True,
+        "fitKa": False,
+        "ab_types": HIgGs,
+    }
+    sample_dist, ag_dist = bootstrap(data, numResample=10, norm="max", **opt_kwargs)
+    return plot_deviation_heatmap(ag_dist[0], ag_dist[1], HIgGs, data.Antigen.values)
